refactor: 将内置的 shape 拆开，支持按需引用
