+++ src/geometry/shape/point/hollow.ts
@@ -1,0 +1,24 @@
+import { each } from '@antv/util';
+import { IGroup } from '../../../dependents';
+import { ShapeInfo, ShapeMarkerCfg } from '../../../interface';
+import { registerShape } from '../base';
+import { drawPoints, HOLLOW_SHAPES, PointSymbols } from './util';
+
+// 添加 hollowShape
+each(HOLLOW_SHAPES, (shapeName: string) => {
+  registerShape('point', shapeName, {
+    draw(cfg: ShapeInfo, container: IGroup) {
+      return drawPoints(this, cfg, container, shapeName, true);
+    },
+    getMarker(markerCfg: ShapeMarkerCfg) {
+      const { color } = markerCfg;
+      return {
+        symbol: PointSymbols[shapeName],
+        style: {
+          r: 4.5,
+          stroke: color,
+        },
+      };
+    },
+  });
+});
