+++ tests/unit/geometry/shape/polygon-spec.ts
@@ -87,50 +87,4 @@ describe('Point shapes', () => {
       });
     });
   });
-
-  describe('hollow', () => {
-    it('draw', () => {
-      const points = PolygonShapeFactory.getShapePoints('hollow', {
-        x: [0.1, 0.3, 0.3, 0.4],
-        y: [0.2, 0.5, 0.12, 0.88],
-      });
-      const shape = PolygonShapeFactory.drawShape(
-        'hollow',
-        {
-          x: 100,
-          y: 100,
-          points,
-          color: 'red',
-          defaultStyle: {
-            ...Theme.geometries.polygon.polygon.default,
-          },
-        },
-        element.container
-      );
-
-      expect(shape.attr('stroke')).toBe('red');
-      expect(shape.attr('path')).toEqual([
-        ['M', 50, 400],
-        ['L', 150, 250],
-        ['L', 150, 440],
-        ['L', 200, 60],
-        ['L', 50, 400],
-        ['Z'],
-      ]);
-    });
-
-    it('getMarker', () => {
-      const markerCfg = PolygonShapeFactory.getMarker('hollow', { color: 'red', isInPolar: false });
-      expect(markerCfg).toEqual({
-        symbol: 'square',
-        style: {
-          r: 4,
-          stroke: 'red',
-          fill: '#fff',
-          fillOpacity: 0,
-          lineWidth: 2,
-        },
-      });
-    });
-  });
 });
